// src/candidates/dto/create-candidate.dto.ts
export class CreateCandidateDto {
  idPerson: number;
  birthDate: Date;
  idCityBorn: number;
  idCityLive: number;
  address: string;
  latitude: number;
  longitude: number;
  idMaritalStatus: number;
  idWorkSchedule: number;
  childrenNo: number;
  childrenDescription: string;
  canRead: boolean;
  canWrite: boolean;
  yoe: number;
  yoeDescription: string;
  cookLevel: number;
  handWash: boolean;
  washingMachine: boolean;
  iron: boolean;
  stove: boolean;
  coffeeMaker: boolean;
  microwave: boolean;
  blender: boolean;
  childrenCareExperience: boolean;
  childrenCareDescription: string;
  tattoos: boolean;
  pregnant: boolean;
  disability: boolean;
  disabilityDescription: string;
  petCare: boolean;
  uniform: boolean;
  availability: number;
  expectedSalary: number;
}